requirements.txt
streamlit~=1.44.1
pandas~=2.2.3
python-dotenv~=1.1.0
google-generativeai

protobuf~=5.29.4